Move ActivityStreams class to separate module
