+++ islandora-connector-derivative/build.gradle
@@ -3,15 +3,17 @@ apply plugin: 'osgi'
 description = 'Islandora 8 Derivative Connector'
 
 dependencies {
-    compile group: 'commons-io', name: 'commons-io', version: commonsIoVersion 
-    compile group: 'org.apache.camel', name: 'camel-blueprint', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-core', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-jackson', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-http4', version: camelVersion
-    compile group: 'org.slf4j', name: 'slf4j-api', version: slf4jVersion
-    testCompile group: 'org.apache.camel', name: 'camel-test-blueprint', version: camelVersion
-    testCompile group: 'org.apache.servicemix.bundles', name: 'org.apache.servicemix.bundles.xerces', version: '2.11.0_1'
-    testCompile group: 'org.ow2.asm', name: 'asm-commons', version: '5.0.3'
+    implementation group: 'commons-io', name: 'commons-io', version: versions.commonsIo 
+    implementation group: 'org.apache.camel', name: 'camel-blueprint', version: versions.camel
+    implementation group: 'org.apache.camel', name: 'camel-core', version: versions.camel
+    implementation group: 'org.apache.camel', name: 'camel-jackson', version: versions.camel
+    implementation group: 'org.apache.camel', name: 'camel-http4', version: versions.camel
+    implementation group: 'org.slf4j', name: 'slf4j-api', version: versions.slf4j
+    implementation project(':islandora-event-support')
+    testImplementation group: 'org.apache.camel', name: 'camel-test-blueprint', version: versions.camel
+    testImplementation group: 'org.apache.servicemix.bundles', name: 'org.apache.servicemix.bundles.xerces', version: versions.xercesServiceMix
+    testImplementation group: 'org.ow2.asm', name: 'asm-commons', version: versions.asmCommons
+    testRuntimeOnly group: 'org.slf4j', name: 'slf4j-simple', version: versions.slf4j
 }
 
 jar {
@@ -22,8 +24,19 @@ jar {
       license project.license
 
       instruction 'Import-Package', 'org.apache.camel.component.http4,' +
-                            "org.apache.camel;version=\"${camelVersion}\"," +
+                            "org.apache.camel;version=\"${versions.camel}\"," +
                             defaultOsgiImports
       instruction 'Export-Package', 'ca.islandora.alpaca.connector.derivative'
+    }
+}
+
+test {
+    testLogging {
+        // Make sure output from standard out or error is shown
+        // in Gradle output.
+        // Uncomment the below 3 lines while debugging.
+        // events 'standard_out', 'standard_error'
+        // exceptionFormat = 'full'
+        // displayGranularity = 0
     }
 }
