+++ islandora-event-support/build.gradle
@@ -1,0 +1,21 @@
+apply plugin: 'osgi'
+
+description = 'Islandora 8 Supporting Libraries'
+
+dependencies {
+    implementation group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: versions.jackson
+    testImplementation group: 'junit', name: 'junit', version: versions.junit
+}
+
+jar {
+    manifest {
+      description project.description
+      docURL project.docURL
+      vendor project.vendor
+      license project.license
+      
+      instruction 'Import-Package', 'com.fasterxml.jackson.annotations,' +
+                            defaultOsgiImports
+      instruction 'Export-Package', 'ca.islandora.alpaca.support'
+    }
+}
