+++ islandora-event-support/src/main/java/ca/islandora/alpaca/support/event/AS2Object.java
@@ -16,7 +16,9 @@
  * limitations under the License.
  */
 
-package ca.islandora.alpaca.connector.derivative.event;
+package ca.islandora.alpaca.support.event;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
  * POJO for a user performing an action.  Part of a AS2Event.
@@ -67,7 +69,23 @@ public class AS2Object {
         this.url = url;
     }
 
+    /**
+     * @return true or false
+     */
+    @JsonProperty(value = "isNewVersion")
+    public Boolean getIsNewVersion() {
+        return isNewVersion;
+    }
+
+    /**
+     * @param isNewVersion true or false
+     */
+    public void setIsNewVersion(final Boolean isNewVersion) {
+        this.isNewVersion = isNewVersion;
+    }
+
     private String type;
     private String id;
     private AS2Url[] url;
+    private Boolean isNewVersion;
 }
