+++ islandora-http-client/build.gradle
@@ -3,9 +3,8 @@ apply plugin: 'osgi'
 description = 'Islandora 8 HTTP Client'
 
 dependencies {
-    compile group: 'org.apache.httpcomponents', name: 'httpclient-osgi', version: '4.5.3'
-
-    testCompile group: 'junit', name: 'junit', version:'4.12'
+    implementation group: 'org.apache.httpcomponents', name: 'httpclient-osgi', version: versions.httpClientOsgi
+    testImplementation group: 'junit', name: 'junit', version: versions.junit
 }
 
 jar {
