+++ islandora-indexing-fcrepo/build.gradle
@@ -3,17 +3,19 @@ apply plugin: 'osgi'
 description = 'Islandora 8 Fcrepo Indexer'
 
 dependencies {
-    compile group: 'org.apache.camel', name: 'camel-core', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-blueprint', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-http4', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-jackson', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-jsonpath', version: camelVersion
-    compile group: 'org.slf4j', name: 'slf4j-api', version: slf4jVersion
-    compile group: 'commons-io', name: 'commons-io', version: commonsIoVersion 
-    testCompile group: 'org.apache.camel', name: 'camel-test-blueprint', version: camelVersion
-    testCompile group: 'org.apache.servicemix.bundles', name: 'org.apache.servicemix.bundles.xerces', version: '2.11.0_1'
-    testCompile group: 'org.ow2.asm', name: 'asm-commons', version: '5.0.3'
-    testCompile group: 'com.googlecode.junit-toolbox', name: 'junit-toolbox', version: '2.3'
+    implementation group: 'org.apache.camel', name: 'camel-core', version: versions.camel
+    implementation group: 'org.apache.camel', name: 'camel-blueprint', version: versions.camel
+    implementation group: 'org.apache.camel', name: 'camel-http4', version: versions.camel
+    implementation group: 'org.apache.camel', name: 'camel-jackson', version: versions.camel
+    implementation group: 'org.apache.camel', name: 'camel-jsonpath', version: versions.camel
+    implementation project(':islandora-event-support')
+    implementation group: 'org.slf4j', name: 'slf4j-api', version: versions.slf4j
+    implementation group: 'commons-io', name: 'commons-io', version: versions.commonsIo 
+    testImplementation group: 'org.apache.camel', name: 'camel-test-blueprint', version: versions.camel
+    testImplementation group: 'org.apache.servicemix.bundles', name: 'org.apache.servicemix.bundles.xerces', version: versions.xercesServiceMix
+    testImplementation group: 'org.ow2.asm', name: 'asm-commons', version: versions.asmCommons
+    testImplementation group: 'com.googlecode.junit-toolbox', name: 'junit-toolbox', version: versions.junitToolbox
+    testRuntimeOnly group: 'org.slf4j', name: 'slf4j-simple', version: versions.slf4j
 }
 
 test {
@@ -29,7 +31,8 @@ jar {
 
       instruction 'Import-Package', 'org.apache.camel.component.http4,' +
                             "org.apache.camel.jsonpath," +
-                            "org.apache.camel;version=\"${camelVersion}\"," +
+                            "org.apache.camel;version=\"${versions.camel}\"," +
+                            "ca.islandora.alpaca.support," +
                             defaultOsgiImports
       instruction 'Export-Package', 'ca.islandora.alpaca.indexing.fcrepo'
     }
