+++ islandora-indexing-triplestore/build.gradle
@@ -3,15 +3,16 @@ apply plugin: 'osgi'
 description = 'Islandora 8 Triplestore Indexer'
 
 dependencies {
-    compile group: 'org.apache.camel', name: 'camel-core', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-blueprint', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-http4', version: camelVersion
-    compile group: 'org.apache.camel', name: 'camel-jsonpath', version: camelVersion
-    compile group: 'org.fcrepo.camel', name: 'fcrepo-camel', version: fcrepoCamelVersion
-    compile group: 'org.slf4j', name: 'slf4j-api', version: slf4jVersion
-    testCompile group: 'org.apache.camel', name: 'camel-test-blueprint', version: camelVersion
-    testCompile group: 'org.apache.servicemix.bundles', name: 'org.apache.servicemix.bundles.xerces', version: '2.11.0_1'
-    testCompile group: 'org.ow2.asm', name: 'asm-commons', version: '5.0.3'
+    implementation group: 'org.apache.camel', name: 'camel-core', version: versions.camel
+    implementation group: 'org.apache.camel', name: 'camel-blueprint', version: versions.camel
+    implementation group: 'org.apache.camel', name: 'camel-http4', version: versions.camel
+    implementation group: 'org.apache.camel', name: 'camel-jsonpath', version: versions.camel
+    implementation group: 'org.fcrepo.camel', name: 'fcrepo-camel', version: versions.fcrepoCamel
+    implementation group: 'org.slf4j', name: 'slf4j-api', version: versions.slf4j
+    testImplementation group: 'org.apache.camel', name: 'camel-test-blueprint', version: versions.camel
+    testImplementation group: 'org.apache.servicemix.bundles', name: 'org.apache.servicemix.bundles.xerces', version: versions.xercesServiceMix
+    testImplementation group: 'org.ow2.asm', name: 'asm-commons', version: versions.asmCommons
+    testRuntimeOnly group: 'org.slf4j', name: 'slf4j-simple', version: versions.slf4j
 }
 
 jar {
@@ -22,8 +23,8 @@ jar {
       license project.license
 
       instruction 'Import-Package', 'org.apache.camel.component.http4,' +
-                            "org.apache.camel;version=\"${camelVersion}\"," +
-                            "org.fcrepo.camel.processor;version=\"${fcrepoCamelVersion}\"," +
+                            "org.apache.camel;version=\"${versions.camel}\"," +
+                            "org.fcrepo.camel.processor;version=\"${versions.fcrepoCamel}\"," +
                             defaultOsgiImports
       instruction 'Export-Package', 'ca.islandora.alpaca.indexing.triplestore'
     }
