+++ settings.gradle
@@ -1,10 +1,12 @@
 include ':islandora-karaf'
+include ':islandora-event-support'
 include ':islandora-indexing-triplestore'
 include ':islandora-indexing-fcrepo'
 include ':islandora-connector-derivative'
 include ':islandora-http-client'
 
 project(':islandora-karaf').projectDir = "$rootDir/karaf" as File
+project(':islandora-event-support').projectDir = "$rootDir/islandora-event-support" as File
 project(':islandora-indexing-triplestore').projectDir = "$rootDir/islandora-indexing-triplestore" as File
 project(':islandora-indexing-fcrepo').projectDir = "$rootDir/islandora-indexing-fcrepo" as File
 project(':islandora-connector-derivative').projectDir = "$rootDir/islandora-connector-derivative" as File
